requirements.txt
# Core dependencies for Connectivity Monitor
# Author: Mason Gonzalez
# Repository: https://github.com/Kernal-mason/gl.itnet-crash-reboot-alert

requests>=2.31.0      # For HTTP requests
python-dotenv>=1.0.0  # For environment variable management
